V0.026
added items and the ability to generate an Item object using an id (attributes determined by json data in item_data.json

debug commands for item testing work as of this commit (test-item and test-set-item), and send a message on discord with the items address. Further testing should be done in later versions.